requirements.txt
tensorflow
pandas
numpy 
scikit-learn
tensorboard
matplotlib
streamlit
sklearn.preprocessing